solidified code into its own file
